+++ src/components/signup/SignUpForm.tsx
@@ -1,9 +1,9 @@
 "use client"
 import React, { useState, useEffect } from "react"
 import Button from "@components/common/Button"
-import InputWithLabel from "./InputWithLabel"
+import InputWithLabel from "./InputLabel"
 import GenderButton from "@components/common/GenderButton"
-import { UserValue, SignFormProps } from "@utils/interface/SignUpInterface"
+import { UserValue, SignFormProps } from "@utils/interface/signUpInterface"
 import { registerUser } from "@utils/api"
 import { useForm } from "react-hook-form"
 import { idPattern, passwordPattern, emailPattern, phoneNumberPattern, genderType, usernamePattern } from "./util"
