+++ src/utils/interface/index.ts
@@ -1,3 +1,3 @@
-export type { TextElementProps } from "./TextElementProps"
-export type { FeedListProps, TravelListProps } from "./FeedListProps"
-export type { SignFormProps, InputProps, BirthDateSelectorProps } from "./SignUpInterface"
+export type { TextElementProps } from "./textElementProps"
+export type { FeedListProps, TravelListProps } from "./feedListProps"
+export type { SignFormProps, InputProps, BirthDateSelectorProps } from "./signUpInterface"
